+++ R/stat_compare_means.R
@@ -37,6 +37,8 @@ NULL
 #'@param label.x,label.y \code{numeric} Coordinates (in data units) to be used
 #'  for absolute positioning of the label. If too short they will be recycled.
 #'@param bracket.size Width of the lines of the bracket.
+#'@param step.increase numeric vector with the increase in fraction of total
+#'  height for every additional comparison to minimize overlap.
 #'@param ... other arguments to pass to \code{\link[ggplot2]{geom_text}} or
 #'  \code{\link[ggplot2]{geom_label}}.
 #'@param na.rm If FALSE (the default), removes missing values with a warning. If
@@ -99,7 +101,7 @@ stat_compare_means <- function(mapping = NULL, data = NULL,
                      comparisons = NULL, hide.ns = FALSE, label.sep = ", ",
                      label = NULL, label.x.npc = "left", label.y.npc = "top",
                      label.x = NULL, label.y = NULL, tip.length = 0.03,
-                     bracket.size = 0.3,
+                     bracket.size = 0.3, step.increase = 0,
                      symnum.args = list(),
                      geom = "text", position = "identity",  na.rm = FALSE, show.legend = NA,
                     inherit.aes = TRUE, ...) {
@@ -135,10 +137,12 @@ stat_compare_means <- function(mapping = NULL, data = NULL,
       if(hide.ns) map_signif_level <- .hide_ns(map_signif_level)
     }
 
-    step_increase <- ifelse(is.null(label.y), 0.12, 0)
+    if(missing(step.increase)){
+      step.increase <- ifelse(is.null(label.y), 0.12, 0)
+    }
     ggsignif::geom_signif(comparisons = comparisons, y_position = label.y,
                           test = method, test.args = method.args,
-                          step_increase = step_increase, size = bracket.size, textsize = size, color = color,
+                          step_increase = step.increase, size = bracket.size, textsize = size, color = color,
                           map_signif_level = map_signif_level, tip_length = tip.length,
                           data = data)
   }